+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["crates/datastore", "crates/audit", "crates/ingress-rpc", "crates/maintenance", "crates/ingress-writer"]
+members = ["crates/datastore", "crates/audit", "crates/ingress-rpc", "crates/maintenance", "crates/ingress-writer", "crates/tracing"]
 resolver = "2"
 
 [workspace.dependencies]
@@ -7,6 +7,7 @@ tips-datastore = { path = "crates/datastore" }
 tips-audit = { path = "crates/audit" }
 tips-maintenance = { path = "crates/maintenance" }
 tips-ingress-writer = { path = "crates/ingress-writer" }
+tips-tracing = { path = "crates/tracing" }
 
 
 # Reth
