+++ crates/ingress-rpc/Cargo.toml
@@ -10,6 +10,7 @@ path = "src/main.rs"
 [dependencies]
 tips-datastore.workspace = true
 tips-audit.workspace = true
+tips-tracing.workspace = true
 jsonrpsee.workspace = true
 alloy-rpc-types-mev.workspace = true
 alloy-primitives.workspace = true
@@ -36,8 +37,4 @@ revm-context-interface.workspace = true
 alloy-signer-local.workspace = true
 reth-optimism-evm.workspace = true
 reth-errors.workspace = true
-opentelemetry.workspace = true
-opentelemetry-otlp.workspace = true
-opentelemetry_sdk.workspace = true
-tracing-opentelemetry.workspace = true
